requirements.txt
pathlib
requests
pendulum
pandas
numpy
datetime
sqlalchemy
pyarrow
python-dotenv
fastparquet
psycopg2-binary
apache-airflow-providers-postgres